requirements.txt
# requirements.txt
Flask==2.2.5
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
WTForms==3.0.1
Werkzeug<2.3.0